+++ environment.yml
@@ -1,0 +1,149 @@
+name: indica_dev
+channels:
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - _openmp_mutex=5.1=1_gnu
+  - ca-certificates=2024.3.11=h06a4308_0
+  - ld_impl_linux-64=2.38=h1181459_1
+  - libffi=3.4.4=h6a678d5_1
+  - libgcc-ng=11.2.0=h1234567_1
+  - libgomp=11.2.0=h1234567_1
+  - libstdcxx-ng=11.2.0=h1234567_1
+  - ncurses=6.4=h6a678d5_0
+  - openssl=3.0.13=h7f8727e_1
+  - pip=24.0=py39h06a4308_0
+  - python=3.9.19=h955ad1f_1
+  - readline=8.2=h5eee18b_0
+  - setuptools=69.5.1=py39h06a4308_0
+  - sqlite=3.45.3=h5eee18b_0
+  - tk=8.6.14=h39e8969_0
+  - wheel=0.43.0=py39h06a4308_0
+  - xz=5.4.6=h5eee18b_1
+  - zlib=1.2.13=h5eee18b_1
+  - pip:
+    - alabaster==0.7.13
+    - aniso8601==9.0.1
+    - attrs==23.2.0
+    - babel==2.15.0
+    - backports-tarfile==1.1.1
+    - blinker==1.8.2
+    - build==1.2.1
+    - cachecontrol==0.14.0
+    - cachy==0.3.0
+    - certifi==2024.2.2
+    - cffi==1.16.0
+    - cfgv==3.4.0
+    - cftime==1.6.3
+    - charset-normalizer==3.3.2
+    - cleo==2.1.0
+    - click==8.1.7
+    - clikit==0.6.2
+    - contourpy==1.1.1
+    - coverage==5.5
+    - crashtest==0.4.1
+    - cryptography==42.0.7
+    - cycler==0.12.1
+    - distlib==0.3.8
+    - docutils==0.17.1
+    - dulwich==0.21.7
+    - execnet==2.1.1
+    - fastjsonschema==2.19.1
+    - filelock==3.14.0
+    - flask==3.0.3
+    - flask-restful==0.3.10
+    - fonttools==4.51.0
+    - html5lib==1.1
+    - hypothesis==5.49.0
+    - identify==2.5.36
+    - idna==3.7
+    - imagesize==1.4.1
+    - importlib-metadata==7.1.0
+    - importlib-resources==6.4.0
+    - indica==0.1.0
+    - iniconfig==2.0.0
+    - installer==0.7.0
+    - isodate==0.6.1
+    - itsdangerous==2.2.0
+    - jaraco-classes==3.4.0
+    - jaraco-context==5.3.0
+    - jaraco-functools==4.0.1
+    - jeepney==0.8.0
+    - jinja2==3.1.4
+    - keyring==24.3.1
+    - kiwisolver==1.4.5
+    - lockfile==0.12.2
+    - lxml==5.2.2
+    - markupsafe==2.1.5
+    - matplotlib==3.7.5
+    - mdsplus==7.49.3
+    - more-itertools==10.2.0
+    - msgpack==1.0.8
+    - mypy==1.10.0
+    - mypy-extensions==1.0.0
+    - netcdf4==1.6.5
+    - networkx==3.1
+    - nodeenv==1.8.0
+    - numpy==1.24.4
+    - packaging==24.0
+    - pandas==2.0.3
+    - pastel==0.2.1
+    - pexpect==4.9.0
+    - pillow==10.3.0
+    - pkginfo==1.10.0
+    - platformdirs==4.2.2
+    - pluggy==1.5.0
+    - poetry==1.8.3
+    - poetry-core==1.9.0
+    - poetry-plugin-export==1.8.0
+    - pre-commit==2.21.0
+    - prov==1.5.3
+    - ptyprocess==0.7.0
+    - py==1.11.0
+    - pycparser==2.22
+    - pygments==2.18.0
+    - pylev==1.4.0
+    - pyparsing==3.1.2
+    - pyproject-hooks==1.1.0
+    - pytest==6.2.5
+    - pytest-cov==2.12.1
+    - pytest-forked==1.6.0
+    - pytest-xdist==2.5.0
+    - python-dateutil==2.9.0.post0
+    - pytz==2024.1
+    - pyyaml==6.0.1
+    - rapidfuzz==3.9.0
+    - rdflib==6.3.2
+    - requests==2.31.0
+    - requests-toolbelt==1.0.0
+    - sal==1.2.5
+    - scipy==1.10.1
+    - secretstorage==3.3.3
+    - shellingham==1.5.4
+    - six==1.16.0
+    - snowballstemmer==2.2.0
+    - sortedcontainers==2.4.0
+    - sphinx==4.5.0
+    - sphinx-rtd-theme==1.3.0
+    - sphinxcontrib-applehelp==1.0.4
+    - sphinxcontrib-devhelp==1.0.2
+    - sphinxcontrib-htmlhelp==2.0.1
+    - sphinxcontrib-jquery==4.1
+    - sphinxcontrib-jsmath==1.0.1
+    - sphinxcontrib-plantuml==0.25
+    - sphinxcontrib-qthelp==1.0.3
+    - sphinxcontrib-serializinghtml==1.1.5
+    - toml==0.10.2
+    - tomli==2.0.1
+    - tomlkit==0.12.5
+    - trove-classifiers==2024.5.17
+    - types-setuptools==57.4.18
+    - typing-extensions==4.11.0
+    - tzdata==2024.1
+    - urllib3==2.2.1
+    - virtualenv==20.26.2
+    - webencodings==0.5.1
+    - werkzeug==3.0.3
+    - xarray==2023.1.0
+    - zipp==3.18.2
+prefix: /home/jussi.hakosalo/.conda/envs/indica_dev
