+++ indica/converters/transect.py
@@ -5,6 +5,7 @@ import numpy as np
 import xarray as xr
 from xarray import DataArray
 
+from indica.utilities import format_coord
 from .abstractconverter import CoordinateTransform
 from ..numpy_typing import Coordinates
 from ..numpy_typing import LabeledArray
@@ -61,15 +62,42 @@ class TransectCoordinates(CoordinateTransform):
         self.name: str = f"{name}_transect_transform"
 
         x1 = np.arange(len(x_positions))
-        self.x1: DataArray = DataArray(x1, coords=[(self.x1_name, x1)])
+        self.x1: DataArray = format_coord(x1, self.x1_name)
         self.x2: DataArray = DataArray(None)
 
         # TODO: add intersection with first walls to restrict possible coordinates
         self._machine_dims = machine_dimensions
+        self.x = xr.DataArray(
+            x_positions,
+            coords={
+                self.x1_name: self.x1.data
+                if isinstance(self.x1, xr.DataArray)
+                else self.x1
+            },
+            dims=self.x1_name,
+        )
+        self.y = xr.DataArray(
+            y_positions,
+            coords={
+                self.x1_name: self.x1.data
+                if isinstance(self.x1, xr.DataArray)
+                else self.x1
+            },
+            dims=self.x1.name,
+        )
+        self.z = xr.DataArray(
+            z_positions,
+            coords={
+                self.x1_name: self.x1.data
+                if isinstance(self.x1, xr.DataArray)
+                else self.x1
+            },
+            dims=self.x1_name,
+        )
 
-        self.x: DataArray = DataArray(x_positions, coords=[(self.x1_name, self.x1)])
-        self.y: DataArray = DataArray(y_positions, coords=[(self.x1_name, self.x1)])
-        self.z: DataArray = DataArray(z_positions, coords=[(self.x1_name, self.x1)])
+        # self.x: DataArray = DataArray(x_positions, coords=[(self.x1_name, self.x1)])
+        # self.y: DataArray = DataArray(y_positions, coords=[(self.x1_name, self.x1)])
+        # self.z: DataArray = DataArray(z_positions, coords=[(self.x1_name, self.x1)])
         self.R: DataArray = np.sqrt(self.x**2 + self.y**2)
         self.rho: DataArray
         self.theta: DataArray
