+++ indica/readers/read_st40.py
@@ -16,9 +16,9 @@ from indica.readers import ST40Reader
 
 INSTRUMENTS: list = [
     "efit",
-    "lines",
+    # "lines",
     "nirh1",
-    "smmh1",
+    "smmh",
     "smmh",
     "cxff_pi",
     "cxff_tws_c",
@@ -368,13 +368,14 @@ def bin_data_in_time(
 
             if "t" in data_quant.coords:
                 data_quant = convert_in_time_dt(tstart, tend, dt, data_quant)
+
             # Using groupedby_bins always removes error from coords so adding it back
             if "error" in raw_data[instr][quant].coords:
                 error = convert_in_time_dt(
                     tstart, tend, dt, raw_data[instr][quant].error
                 )
                 data_quant = data_quant.assign_coords(
-                    error=(raw_data[instr][quant].dims, error)
+                    error=(raw_data[instr][quant].dims, error.data)
                 )
             binned_quantities[quant] = data_quant
         binned_data[instr] = binned_quantities
