+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["poetry-core>=1.3.2"]
+requires = ["poetry-core>=1.7"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.black]
@@ -41,16 +41,16 @@ include = [".git", "indica/git_version"]
 
 [tool.poetry.dependencies]
 corner = "^1.0"
-pandas = "1.3.0"
-python = ">=3.9,<3.11"
+pandas = "2.2.1"
+python = ">=3.11"
 prov = "^1.5.3"
 netCDF4 = "^1.5.4"
-numpy = "1.21.0"
-scipy = "^1.6.2"
-xarray = "0.16.2"
-matplotlib = "^3.3.0"
+numpy = "1.26.4"
+scipy = "^1.9.2"
+xarray = "2024.6.0"
+matplotlib = "^3.8.0"
 sal = "^1.2.2"
-ipython = "^8.8.0"
+ipython = "^8.24.0"
 emcee = "^3.1.4"
 flatdict = "^4.0.1"
 tqdm = "^4.65.0"
@@ -62,7 +62,7 @@ periodictable = "^1.6.1"
 [tool.poetry.dev-dependencies]
 click = "^8.0.0"
 corner = "^1.0"
-coverage = {extras = ["toml"], version = "^5.2.1"}
+coverage = {extras = ["toml"], version = "^7.5.1"}
 hypothesis = {extras = ["numpy"], version = "^5.24.2"}
 mypy = "^0.942"
 pre-commit = "^2.9.2"
